app/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.5
python-dotenv==0.21.1
python-multipart
opencv-python
openai==1.61.1
azure-identity
azure-storage-blob